Share code between functions and arrow functions
